VIH-9307 rename component
