+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/services/hide-components.service.ts
